+++ build_config/test/build_config_test.dart
@@ -14,9 +14,10 @@ void main() {
       'example:a': new BuildTarget(
         builders: {
           'b|b': new TargetBuilderConfig(
+              isEnabled: true,
               generateFor: new InputSet(include: ['lib/a.dart'])),
           'example|h': new TargetBuilderConfig(
-              options: new BuilderOptions({'foo': 'bar'})),
+              isEnabled: true, options: new BuilderOptions({'foo': 'bar'})),
         },
         dependencies: ['b:b', 'c:d'].toSet(),
         package: 'example',
@@ -37,7 +38,7 @@ void main() {
         builderFactories: ['createBuilder'],
         autoApply: AutoApply.dependents,
         isOptional: true,
-        buildTo: BuildTo.cache, // defaulted because of AutoApply.dependents
+        buildTo: BuildTo.cache,
         import: 'package:example/e.dart',
         buildExtensions: {
           '.dart': [
@@ -71,7 +72,7 @@ void main() {
         builderFactories: ['createBuilder'],
         autoApply: AutoApply.none,
         isOptional: false,
-        buildTo: BuildTo.source,
+        buildTo: BuildTo.cache,
         import: 'package:example/builder.dart',
         buildExtensions: {
           '.dart': [