+++ build_runner/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.7.0-dev
+## 0.7.0
 
 ### New Features
 
@@ -21,6 +21,10 @@
 
 - Removed `BuildAction`. Changed `build` and `watch` to take a
   `List<BuilderApplication>`. See `apply` and `applyToRoot` to set these up.
+- Changed `apply` to take a single String argument - a Builder key from
+  `package:build_config` rather than a separate package and builder name.
+- Changed the default value of `hideOutput` from `false` to `true` for `apply`.
+  With `applyToRoot` the value remains `false`.
 - There is now a whitelist of top level directories that will be used as a part
   of the build, and other files will be ignored. For now those directories
   include 'benchmark', 'bin', 'example', 'lib', 'test', 'tool', and 'web'.